Tested out new classification predictors
I wanted to see if there was a non mfcc way to classify the audio. My thought process was to get rid of all audio in the 0 - 1000 kHz range (human speech). Then to perform fft, and pick the top 3 amplitudes from each snippet as predictors. The accuracy was meh, 80%. Either the mfcc's are incredibly overfitting the data, or they actually do a good job (literature suggests that they do a good job). I think the only real way to move forward is to record with the mic we have and go from there because there's quite a big difference between the iphone's mic and the raspberry pi mic.