Adding more classifiers in the biased file
I'm still confused why it's so hard for these models to make the predictions. When I paste the predictors into excel and put on conditional formatting, it's so clear how different the angle grinders are from the environment. Maybe this will improve with better audio data - not completely sure tho... I guess for now we'll stick to mfcc's.